Build 3
I got the Team + Hour log where I originally wanted it, but I need to
add the Daily/Weekly limit

+++ ProjectManagementPrototype/src/ksu/fall2017/swe4663/group1/projectmanagementsystem/team/Person.java
@@ -64,7 +64,7 @@ public class Person implements Serializable {
 		this.team = team;
 	}
 
-	public void reportHours( int duration, WorkedHourType workedHourType )
+	public void reportHours( double duration, WorkedHourType workedHourType )
 			throws PersonNotOnTeamException, InvalidWorkedHourTypeException {
 		LoggingTool.print( "Person: Reporting hours." );
 		if ( this.team == null ) {
